+++ anvil_consortium_manager/adapters/default.py
@@ -2,6 +2,7 @@
 
 from .. import filters, forms, models, tables
 from .account import BaseAccountAdapter
+from .managed_group import BaseManagedGroupAdapter
 from .workspace import BaseWorkspaceAdapter
 
 
@@ -10,6 +11,12 @@ class DefaultAccountAdapter(BaseAccountAdapter):
 
     list_table_class = tables.AccountStaffTable
     list_filterset_class = filters.AccountListFilter
+
+
+class DefaultManagedGroupAdapter(BaseManagedGroupAdapter):
+    """Default adapter to use for ManagedGroups in the app."""
+
+    list_table_class = tables.ManagedGroupStaffTable
 
 
 class DefaultWorkspaceAdapter(BaseWorkspaceAdapter):