+++ anvil_consortium_manager/tests/test_adapters.py
@@ -1,9 +1,12 @@
+import django_tables2
 from django.core.exceptions import ImproperlyConfigured
 from django.forms import Form, ModelForm
 from django.test import TestCase, override_settings
+from django_filters import FilterSet
 
 from ..adapters.account import BaseAccountAdapter
-from ..adapters.default import DefaultAccountAdapter, DefaultWorkspaceAdapter
+from ..adapters.default import DefaultAccountAdapter, DefaultManagedGroupAdapter, DefaultWorkspaceAdapter
+from ..adapters.managed_group import BaseManagedGroupAdapter
 from ..adapters.workspace import (
     AdapterAlreadyRegisteredError,
     AdapterNotRegisteredError,
@@ -13,7 +16,12 @@ from ..adapters.workspace import (
 from ..filters import AccountListFilter, BillingProjectListFilter
 from ..forms import DefaultWorkspaceDataForm, WorkspaceForm
 from ..models import Account, DefaultWorkspaceData
-from ..tables import AccountStaffTable, WorkspaceStaffTable, WorkspaceUserTable
+from ..tables import (
+    AccountStaffTable,
+    ManagedGroupStaffTable,
+    WorkspaceStaffTable,
+    WorkspaceUserTable,
+)
 from . import factories
 from .test_app import filters, forms, models, tables
 from .test_app.adapters import TestWorkspaceAdapter
@@ -47,6 +55,20 @@ class AccountAdapterTestCase(TestCase):
         setattr(TestAdapter, "list_table_class", None)
         with self.assertRaises(ImproperlyConfigured):
             TestAdapter().get_list_table_class()
+
+    def test_list_table_class_wrong_model(self):
+        """get_list_table_class raises ImproperlyConfigured when incorrect model is used for the table."""
+
+        class TestTable(django_tables2.Table):
+            class Meta:
+                model = DefaultWorkspaceData
+                fields = ("name",)
+
+        TestAdapter = self.get_test_adapter()
+        setattr(TestAdapter, "list_table_class", TestTable)
+        with self.assertRaises(ImproperlyConfigured) as e:
+            TestAdapter().get_list_table_class()
+        self.assertIn("anvil_consortium_manager.models.Account", str(e.exception))
 
     def test_list_filterset_class_default(self):
         """get_list_filterset_class returns the correct filter when using the default adapter."""
@@ -82,6 +104,20 @@ class AccountAdapterTestCase(TestCase):
         setattr(TestAdapter, "list_filterset_class", Foo)
         with self.assertRaises(ImproperlyConfigured):
             TestAdapter().get_list_filterset_class()
+
+    def test_list_filterset_class_wrong_model(self):
+        """Raises ImproperlyConfigured when incorrect model is used for the filterset."""
+
+        class TestFilterSet(FilterSet):
+            class Meta:
+                model = models.ManagedGroup
+                fields = {"email": ["icontains"]}
+
+        TestAdapter = self.get_test_adapter()
+        setattr(TestAdapter, "list_filterset_class", TestFilterSet)
+        with self.assertRaises(ImproperlyConfigured) as e:
+            TestAdapter().get_list_filterset_class()
+        self.assertIn("anvil_consortium_manager.models.Account", str(e.exception))
 
     def test_get_autocomplete_queryset_default(self):
         """get_autocomplete_queryset returns the correct queryset when using the default adapter."""
@@ -125,6 +161,50 @@ class AccountAdapterTestCase(TestCase):
         self.assertEqual(TestAdapter().get_autocomplete_label(account), "testuser")
 
 
+class ManagedGroupAdapterTest(TestCase):
+    """Tests for ManagedGroup adapters."""
+
+    def get_test_adapter(self):
+        """Return a test adapter class for use in tests."""
+
+        class TestAdapter(BaseManagedGroupAdapter):
+            list_table_class = tables.TestManagedGroupTable
+
+        return TestAdapter
+
+    def test_list_table_class_default(self):
+        """get_list_table_class returns the correct table when using the default adapter."""
+        self.assertEqual(DefaultManagedGroupAdapter().get_list_table_class(), ManagedGroupStaffTable)
+
+    def test_list_table_class_custom(self):
+        """get_list_table_class returns the correct table when using a custom adapter."""
+        TestAdapter = self.get_test_adapter()
+        setattr(TestAdapter, "list_table_class", tables.TestManagedGroupTable)
+        self.assertEqual(TestAdapter().get_list_table_class(), tables.TestManagedGroupTable)
+
+    def test_list_table_class_none(self):
+        """get_list_table_class raises ImproperlyConfigured when list_table_class is not set."""
+        TestAdapter = self.get_test_adapter()
+        setattr(TestAdapter, "list_table_class", None)
+        with self.assertRaises(ImproperlyConfigured) as e:
+            TestAdapter().get_list_table_class()
+        self.assertIn("Set `list_table_class`", str(e.exception))
+
+    def test_list_table_class_wrong_model(self):
+        """get_list_table_class raises ImproperlyConfigured when incorrect model is used for the table."""
+
+        class TestTable(django_tables2.Table):
+            class Meta:
+                model = Account
+                fields = ("email",)
+
+        TestAdapter = self.get_test_adapter()
+        setattr(TestAdapter, "list_table_class", TestTable)
+        with self.assertRaises(ImproperlyConfigured) as e:
+            TestAdapter().get_list_table_class()
+        self.assertIn("anvil_consortium_manager.models.ManagedGroup", str(e.exception))
+
+
 class WorkspaceAdapterTest(TestCase):
     """Tests for Workspace adapters."""
 
@@ -161,6 +241,20 @@ class WorkspaceAdapterTest(TestCase):
         with self.assertRaises(ImproperlyConfigured):
             TestAdapter().get_list_table_class_staff_view()
 
+    def test_list_table_class_staff_view_wrong_model(self):
+        """get_list_table_class_staff_view raises ImproperlyConfigured when incorrect model is used for the table."""
+
+        class TestTable(django_tables2.Table):
+            class Meta:
+                model = Account
+                fields = ("email",)
+
+        TestAdapter = self.get_test_adapter()
+        setattr(TestAdapter, "list_table_class_staff_view", TestTable)
+        with self.assertRaises(ImproperlyConfigured) as e:
+            TestAdapter().get_list_table_class_staff_view()
+        self.assertIn("anvil_consortium_manager.models.Workspace", str(e.exception))
+
     def test_list_table_class_view_default(self):
         """get_list_table_class returns the correct table when using the default adapter."""
         self.assertEqual(DefaultWorkspaceAdapter().get_list_table_class_view(), WorkspaceUserTable)
@@ -177,6 +271,20 @@ class WorkspaceAdapterTest(TestCase):
         setattr(TestAdapter, "list_table_class_view", None)
         with self.assertRaises(ImproperlyConfigured):
             TestAdapter().get_list_table_class_view()
+
+    def test_list_table_class_view_wrong_model(self):
+        """get_list_table_class_view raises ImproperlyConfigured when incorrect model is used for the table."""
+
+        class TestTable(django_tables2.Table):
+            class Meta:
+                model = Account
+                fields = ("email",)
+
+        TestAdapter = self.get_test_adapter()
+        setattr(TestAdapter, "list_table_class_view", TestTable)
+        with self.assertRaises(ImproperlyConfigured) as e:
+            TestAdapter().get_list_table_class_view()
+        self.assertIn("anvil_consortium_manager.models.Workspace", str(e.exception))
 
     def test_get_workspace_form_class_default(self):
         """get_workspace_form_class returns the correct form when using the default adapter."""